集成DAO层代码

+++ src/main/java/com/tudog/graphqldemo01/model/BookInput.java
@@ -6,12 +6,12 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class BookInput{
-    private Integer id;
+    private Long id;
     private String name;
     private Integer pageCount;
     private Integer authorId;
 
-    public BookInput(Integer id, String name, Integer pageCount, Integer authorId) {
+    public BookInput(Long id, String name, Integer pageCount, Integer authorId) {
         this.id = id;
         this.name = name;
         this.pageCount = pageCount;
